Squashed commit of the following:
commit b54ee10c2bf64fabb11466e72b5cf2e224f620a2
Author: default-writer <funcelot@outlook.com>
Date:   Sun Dec 11 05:35:41 2022 +0000

    added bazel dependencies

+++ bin/venv.sh
@@ -1,15 +1,14 @@
 #!/bin/bash -e
 uid=$(id -u)
 
-if [ "${uid}" -eq 0 ]
-then
-  echo "Please run as user"
-  exit
+if [ "${uid}" -eq 0 ]; then
+	echo "Please run as user"
+	exit
 fi
 
 pwd=$(pwd)
 
-${pwd}/bin/clean.sh
+"${pwd}/bin/clean.sh"
 
 rm -rf venv
 
@@ -17,6 +16,6 @@ python3 -m venv venv
 
 . venv/bin/activate
 python3 -m pip install --upgrade pip
-python3 -m pip install -r requirements.txt
+[ -f "requirements.txt" ] && python3 -m pip install -r requirements.txt
 
-cd ${pwd}
+cd "${pwd}"
